+++ src/hooks/useBottomSheet.ts
@@ -1,0 +1,15 @@
+import { useState } from 'react';
+
+export const useBottomSheet = () => {
+  const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false);
+
+  const handleBottomSheetOpen = () => {
+    setIsBottomSheetOpen(true);
+  };
+
+  const handleBottomSheetClose = () => {
+    setIsBottomSheetOpen(false);
+  };
+
+  return { isBottomSheetOpen, handleBottomSheetOpen, handleBottomSheetClose };
+};
